rename DatabaseFunctions.php to UserMethods.php
